+++ src/services-hooks/dashboards/useGetSalesAnalytics.tsx
@@ -20,7 +20,13 @@ const months = [
   "NOV",
   "DEC",
 ];
-export default function useGetSalesAnalytics() {
+export default function useGetSalesAnalytics({
+  start_date,
+  end_date,
+}: {
+  start_date?: string;
+  end_date?: string;
+}) {
   const axios = useAxios(true);
   const dispatch = useAppDispatch();
   const { status, data, monthlyAmountsStatistics } = useAppSelector(
@@ -31,9 +37,12 @@ export default function useGetSalesAnalytics() {
 
   const getSalesAnalytics = useCallback(async () => {
     setIsLoading(true);
+    const currentYear = new Date().getFullYear();
     try {
       const response = await axios.get(
-        "/admin/dashboard/sales-analytics?start_date=2024-01-01&end_date=2024-12-30"
+        start_date && end_date
+          ? `/admin/dashboard/sales-analytics?start_date=${start_date}&end_date=${end_date}`
+          : `/admin/dashboard/sales-analytics?start_date=${currentYear}-01-01&end_date=${currentYear}-12-30`
       );
       const { sales } = response?.data?.data;
       const endResult = months.map((item, index) => {
@@ -48,13 +57,14 @@ export default function useGetSalesAnalytics() {
     } catch (error) {
       setIsFailed(true);
     }
-  }, []);
+  }, [start_date, end_date]);
 
   useEffect(() => {
-    if (!status) {
+    if ((start_date && end_date) || !status) {
+      //check if stored data is used or new request is fetched. Also check start date and end date has been provided to load the new data
       getSalesAnalytics();
     }
-  }, [status]);
+  }, [status, start_date, end_date]);
 
   return {
     data: { data, stats: monthlyAmountsStatistics },